+++ board/yuzukihomekit/usb_test/main.c
@@ -41,7 +41,7 @@ int main(void) {
     printk_info("Hello World!\n");
 
     /* Initialize the DRAM and enable memory management unit (MMU). */
-    uint64_t dram_size = sunxi_dram_init(&dram_para);
+    uint32_t dram_size = sunxi_dram_init(&dram_para);
     arm32_mmu_enable(SDRAM_BASE, dram_size);
 
     /* Debug message to indicate that MMU is enabled. */
